delete xtra

+++ Controllers/AdminController.cs
@@ -199,7 +199,7 @@ namespace ConsoleApplication.Controllers
             bool formatExists = false;
             Items s = itemsRepository.Get(id);
             String format= s.Format;
-            if(!format.Equals("Link"))
+            if(!format.Equals("Link")&&(!String.IsNullOrEmpty(s.Path)))
             {
                 System.IO.File.Delete(s.Path);
